Moving Bizkaia tests to a subfolder + renaming test classes
