+++ src/Basque/Mews.Fiscalizations.Basque.Tests/Bizkaia/Documents/LROE_240_FacturasEmitidasConSG_Correcta.xml
