+++ src/Basque/Mews.Fiscalizations.Basque.Tests/Bizkaia/Xsd/LROE_PJ_240_1_1_FacturasEmitidas_ConSG_AltaRespuesta_V1_0_1.xsd
