+++ src/Basque/Mews.Fiscalizations.Basque.Tests/Bizkaia/Xsd/batuz_TiposBasicos.xsd
