+++ src/Basque/Mews.Fiscalizations.Basque.Tests/Mews.Fiscalizations.Basque.Tests.csproj
@@ -17,36 +17,36 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Documents\" />
-    <Folder Include="Xsd\" />
+    <Folder Include="Bizkaia\Documents\" />
+    <Folder Include="Bizkaia\Xsd\" />
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="Documents\LROE_240_FacturasEmitidasConSG_Correcta.xml">
+    <None Update="Bizkaia\Documents\LROE_240_FacturasEmitidasConSG_Correcta.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Documents\LROE_240_FacturasEmitidasConSG_Incorrecta.xml">
+    <None Update="Bizkaia\Documents\LROE_240_FacturasEmitidasConSG_Incorrecta.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Documents\LROE_240_FacturasEmitidasConSG_Parc_Correcta.xml">
+    <None Update="Bizkaia\Documents\LROE_240_FacturasEmitidasConSG_Parc_Correcta.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Documents\ticketBai.xml">
+    <None Update="Bizkaia\Documents\ticketBai.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Xsd\batuz_Enumerados.xsd">
+    <None Update="Bizkaia\Xsd\batuz_Enumerados.xsd">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Xsd\batuz_TiposBasicos.xsd">
+    <None Update="Bizkaia\Xsd\batuz_TiposBasicos.xsd">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Xsd\batuz_TiposComplejos.xsd">
+    <None Update="Bizkaia\Xsd\batuz_TiposComplejos.xsd">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Xsd\LROE_PJ_240_1_1_FacturasEmitidas_ConSG_AltaPeticion_V1_0_2.xsd">
+    <None Update="Bizkaia\Xsd\LROE_PJ_240_1_1_FacturasEmitidas_ConSG_AltaPeticion_V1_0_2.xsd">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Xsd\LROE_PJ_240_1_1_FacturasEmitidas_ConSG_AltaRespuesta_V1_0_1.xsd">
+    <None Update="Bizkaia\Xsd\LROE_PJ_240_1_1_FacturasEmitidas_ConSG_AltaRespuesta_V1_0_1.xsd">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
